Added addToCart method in cart service.
